+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Allow to prevent auto closing via `form.$onSubmit = (e) => e.preventDefault()`
 - **Controls**
   - Add css-animation for box-shadow/borders
+- **Radio**. Add comparison by item.id to allow use cloned items/value
 - **Form**
   - Add `form.$validate(...)` for manual triggering validations
   - Changed spinner to custom friendly looked